% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Ancestry.list}
\alias{Ancestry.list}
\title{Allele ancestries for the American maize NAM dataset}
\format{
A list of three numeric matrices
}
\usage{
Ancestry.list
}
\description{
\code{Ancestry.list} is a list of numeric matrices corresponding to three populations
of the American maize NAM experiment. Each matrix has 0/1 entries representing
the ancestry of alleles (0: homozygous for B73 alleles and 1: homozygous
for alternative parent alleles).
Each matrix includes 191 to 196 recombinant inbred lines characterized at 1,106 markers.
}
\keyword{datasets}
