% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{EpiTest.plot.mean}
\alias{EpiTest.plot.mean}
\title{Display directional epistasis plot}
\usage{
EpiTest.plot.mean(ETest, Title = NULL, Family_names = NULL, Alpha = NULL)
}
\arguments{
\item{ETest}{a list as obtained from the EpiTest.fit function}

\item{Title}{(optional) a character string to be used as a title for the graph}

\item{Family_names}{(optional) a character vector with family names}

\item{Alpha}{type I error for the test}
}
\value{
A ggplot graph
}
\description{
Display directional epistasis plot
}
