% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{EpiTest.plot.var}
\alias{EpiTest.plot.var}
\title{Display variance component graphs}
\usage{
EpiTest.plot.var(ETest, Title = NULL, Family_names = NULL, Colors = NULL)
}
\arguments{
\item{ETest}{a list as obtained from the EpiTest.fit function}

\item{Title}{(optional) a character string to be used as a title for the graph}

\item{Family_names}{(optional) a character vector with family names}

\item{Colors}{(optional) a character vector with three colors for the
graphical display}
}
\value{
a ggplot graph
}
\description{
Display variance component graphs
}
