% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Parents}
\alias{Parents}
\title{Phenotypes for parental lines of the American maize NAM dataset}
\format{
A data.frame
}
\usage{
Parents
}
\description{
\code{Parents} is a data.frame with four rows. The first row corresponds to the
parental  inbred line common to all bi-parental populations of the American maize NAM experiment
(i.e., B73), whereas the three other rows correspond to alternative inbred parental lines used to
generate other bi-parental populations. For each parental inbred line, the phenotypic values
for five traits are provided, along with two extra columns:
Genotype (providing the line ID) and Weight (providing the weight associated
with the individual, see function EpiTest.fit for details).
}
\keyword{datasets}
