% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pchisqmix}
\alias{pchisqmix}
\title{Chi-Squared Mixtures Distribution Function}
\usage{
pchisqmix(quant, s, q, lower.tail = TRUE)
}
\arguments{
\item{quant}{a quantile}

\item{s}{number of fixed effects to be tested}

\item{q}{number of random effects to be tested}

\item{lower.tail}{logical. if TRUE (default), probabilities are \eqn{P[X<=x]}, otherwise \eqn{P[X>x]}.}
}
\value{
a probability.
}
\description{
The approximate null distribution of a likelihood ratio for 2 nested mixed models,
where both fixed and random effects are tested simultaneously, is a very specific mixture
of chi-square distributions. It depends on both the number of random effects
and the number of fixed effects to be tested simultaneously. Note that this function is a copy
of the \code{pchisqmix} function of the \code{TcGSA} package.
}
