% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Pheno.list}
\alias{Pheno.list}
\title{Phenotypes for the American maize NAM dataset}
\format{
A list of three data.frames
}
\usage{
Pheno.list
}
\description{
\code{Pheno.list} is a list of three data.frames corresponding to three
populations of the American maize NAM experiment.
Each data.frame contains the phenotypic values of 191 to 196 recombinant
inbred lines (RILs) evaluated for five different traits, along with
two extra columns:
Genotype (providing the RIL ID) and Weight (providing the weight associated
with the RIL, see function EpiTest.fit for details).
}
\keyword{datasets}
